chore: Movied test redis container into event duplication test
